Refactor create to use sub-routines
