+++ tools/create/src/create.types.ts
@@ -1,6 +1,7 @@
 export interface CreatePackageOptions {
   directory?: string;
   scope?: string;
+  parent?: string;
 }
 
 export interface TemplateParameters {
